# go/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/relationship/friend_request_list": {
            "get": {
                "description": "Provide every name of user request to be your friend",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Relationship"
                ],
                "summary": "Get friend request",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication token",
                        "name": "Token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get Request List successful",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "unexpect Authorization",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    }
}